remove transformer-components/ and move its contents into appropriate places
