+++ src/components/transformer-template/util.ts
@@ -1,0 +1,19 @@
+import { DataSet } from "../../transformers/types";
+import { createTableWithDataSet } from "../../lib/codapPhone";
+
+/**
+ * This function takes a dataset as well as a `doUpdate` flag and either
+ * creates a new table for the dataset or updates an existing one accordingly.
+ *
+ * @returns The name of the newly created context.
+ */
+export async function applyNewDataSet(
+  dataSet: DataSet,
+  name: string | undefined,
+  description: string | undefined
+): Promise<string> {
+  // if doUpdate is true then we should update a previously created table
+  // rather than creating a new one
+  const [newContext] = await createTableWithDataSet(dataSet, name, description);
+  return newContext.name;
+}
