+++ src/transformers/compare.ts
@@ -7,7 +7,7 @@ import {
 } from "./util";
 import { TransformerTemplateState } from "../components/transformer-template/TransformerTemplate";
 import { getContextAndDataSet } from "../lib/codapPhone";
-import { readableName } from "../transformer-components/util";
+import { readableName } from "../transformers/util";
 import { uniqueName } from "../lib/utils/names";
 import {
   colorToRgbString,
