+++ src/transformers/join.ts
@@ -2,7 +2,7 @@ import { DataSet, TransformationOutput } from "./types";
 import { uniqueName } from "../lib/utils/names";
 import { TransformerTemplateState } from "../components/transformer-template/TransformerTemplate";
 import { getContextAndDataSet } from "../lib/codapPhone";
-import { readableName } from "../transformer-components/util";
+import { readableName } from "../transformers/util";
 import {
   shallowCopy,
   cloneCollection,
